Move wirelesstag shared constants to separate module
