+++ homeassistant/components/wirelesstag/binary_sensor.py
@@ -15,12 +15,9 @@ from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from . import (
-    DOMAIN as WIRELESSTAG_DOMAIN,
-    SIGNAL_BINARY_EVENT_UPDATE,
-    WirelessTagBaseSensor,
-    async_migrate_unique_id,
-)
+from . import WirelessTagBaseSensor
+from .const import DOMAIN, SIGNAL_BINARY_EVENT_UPDATE
+from .util import async_migrate_unique_id
 
 # On means in range, Off means out of range
 SENSOR_PRESENCE = "presence"
@@ -84,7 +81,7 @@ async def async_setup_platform(
     discovery_info: DiscoveryInfoType | None = None,
 ) -> None:
     """Set up the platform for a WirelessTags."""
-    platform = hass.data[WIRELESSTAG_DOMAIN]
+    platform = hass.data[DOMAIN]
 
     sensors = []
     tags = platform.tags
