+++ homeassistant/components/wirelesstag/const.py
@@ -1,0 +1,11 @@
+"""Support for Wireless Sensor Tags."""
+
+DOMAIN = "wirelesstag"
+
+# Template for signal - first parameter is tag_id,
+# second, tag manager mac address
+SIGNAL_TAG_UPDATE = "wirelesstag.tag_info_updated_{}_{}"
+
+# Template for signal - tag_id, sensor type and
+# tag manager mac address
+SIGNAL_BINARY_EVENT_UPDATE = "wirelesstag.binary_event_updated_{}_{}_{}"
